refactor -> move CreateRoomForm to package model.form
